requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
pypaystack2==2.0.3
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
